pyproject.toml
[project]
name = "bot-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.22.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "pillow>=12.0.0",
    "python-decouple>=3.8",
]
